Warn when potentially overwrite git folder
Required moving large chunks of controller code regarding installation
status responses to App from ModList.
Also improves UI of overwrite/auto-update modals.
